Colocando o SQL Lite

+++ app/src/main/java/br/com/example/crud/model/User.kt
@@ -1,17 +1,17 @@
 package br.com.example.crud.model
 
-data class User constructor (
-    val name: String,
-    val email: String,
-    val password: String,
-    val isActive: Boolean
-    ){
-    constructor(
-        id: Int? = null,
-        name: String,
-        email: String,
-        password: String,
-        isActive: Boolean
-    ) : this(name, email, password, isActive)
+data class User constructor(
+	val name: String,
+	val email: String,
+	val password: String,
+	val isActive: Boolean
+) {
+	constructor(
+		id: Long? = null,
+		name: String,
+		email: String,
+		password: String,
+		isActive: Boolean
+	) : this(name, email, password, isActive)
 }
